jwt 인가 과정 추가_회원기능

+++ hackathon/src/main/java/org/spring/hackathon/member/service/MemberService.java
@@ -6,6 +6,7 @@ import org.spring.hackathon.member.domain.MemberEntity;
 import org.spring.hackathon.member.dto.MemberDto;
 import org.spring.hackathon.member.repository.MemberImageRepository;
 import org.spring.hackathon.member.repository.MemberRepository;
+import org.spring.hackathon.security.utils.JwtProvider;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -17,25 +18,37 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class MemberService {
 
+  private final JwtProvider jwtProvider;
   private final MemberRepository memberRepository;
   private final MemberImageRepository memberImageRepository;
   private final PasswordEncoder passwordEncoder;
 
   //마이페이지 회원 정보 조회
-  public MemberDto memberMyPageView(Long memberNo) {
+  public MemberDto memberMyPageView(Long memberNo, String token) {
 
-    Optional<MemberEntity> memberEntity = memberRepository.findById(memberNo);
+    Optional<MemberEntity> memberCheck = memberRepository.findById(memberNo);
+    MemberEntity memberEntityGet = memberCheck.get();
 
-    if(memberEntity.get().getMemberAttachImage() == 1) {
+    String memberId = jwtProvider.getUserId(token.substring(7));
+
+    if(!memberCheck.isPresent()) {
+      throw new RuntimeException("정상적인 접근이 아닙니다(회원 확인 불가)");
+    }
+
+    if(!memberEntityGet.getMemberId().equals(memberId)) {
+      throw new RuntimeException("정상적인 접근이 아닙니다(로그인 정보 불일치)");
+    }
+
+    if(memberEntityGet.getMemberAttachImage() == 1) {
 
       String filePath = memberImageRepository.findFilePathWithMemberFk(memberNo);
-      MemberDto memberDto = MemberConstructor.memberEntityToDto(memberEntity.get(), filePath);
+      MemberDto memberDto = MemberConstructor.memberEntityToDto(memberEntityGet, filePath);
 
       return memberDto;
 
     } else {
 
-      MemberDto memberDto = MemberConstructor.memberEntityToDto(memberEntity.get(), null);
+      MemberDto memberDto = MemberConstructor.memberEntityToDto(memberEntityGet, null);
       return memberDto;
 
     }
@@ -44,22 +57,31 @@ public class MemberService {
 
   //회원 정보 수정
   @Transactional
-  public MemberDto memberInfoUpdate(Long memberNo, MemberDto dto) {
+  public MemberDto memberInfoUpdate(Long memberNo, MemberDto dto, String token) {
 
-    Optional<MemberEntity> optionalMemberEntity = memberRepository.findById(memberNo);
+    Optional<MemberEntity> memberCheck = memberRepository.findById(memberNo);
+    MemberEntity memberEntityGet = memberCheck.get();
 
-    if (optionalMemberEntity.isPresent()){
+    String memberId = jwtProvider.getUserId(token.substring(7));
 
-      MemberEntity memberUpdateEntity = optionalMemberEntity.get();
+    if(!memberCheck.isPresent()) {
+      throw new RuntimeException("정상적인 접근이 아닙니다(회원 확인 불가)");
+    }
 
-      memberUpdateEntity.setMemberPassword(passwordEncoder.encode(dto.getMemberPassword()));
-      memberUpdateEntity.setMemberName(dto.getMemberName());
-      memberUpdateEntity.setMemberEmail(dto.getMemberEmail());
-      memberUpdateEntity.setMemberAddress(dto.getMemberAddress());
-      memberUpdateEntity.setMemberIntro(dto.getMemberIntro());
-      memberUpdateEntity.setMemberAttachImage(dto.getMemberAttachImage());
+    if(!memberEntityGet.getMemberId().equals(memberId)) {
+      throw new RuntimeException("정상적인 접근이 아닙니다(로그인 정보 불일치)");
+    }
 
-      memberRepository.save(memberUpdateEntity);
+    if (memberCheck.isPresent()){
+
+      memberEntityGet.setMemberPassword(passwordEncoder.encode(dto.getMemberPassword()));
+      memberEntityGet.setMemberName(dto.getMemberName());
+      memberEntityGet.setMemberEmail(dto.getMemberEmail());
+      memberEntityGet.setMemberAddress(dto.getMemberAddress());
+      memberEntityGet.setMemberIntro(dto.getMemberIntro());
+      memberEntityGet.setMemberAttachImage(dto.getMemberAttachImage());
+
+      memberRepository.save(memberEntityGet);
 
       }
 
@@ -68,7 +90,20 @@ public class MemberService {
     }
    
   //회원 탈퇴
-  public void memberDelete(Long memberNo) {
+  public void memberDelete(Long memberNo, String token) {
+
+    Optional<MemberEntity> memberCheck = memberRepository.findById(memberNo);
+    MemberEntity memberEntityGet = memberCheck.get();
+
+    String memberId = jwtProvider.getUserId(token.substring(7));
+
+    if(!memberCheck.isPresent()) {
+      throw new RuntimeException("정상적인 접근이 아닙니다(회원 확인 불가)");
+    }
+
+    if(!memberEntityGet.getMemberId().equals(memberId)) {
+      throw new RuntimeException("정상적인 접근이 아닙니다(로그인 정보 불일치)");
+    }
 
     memberRepository.deleteById(memberNo);
